+++ src/Fx.Amiya.Background.Api/Vo/HospitalConsulationOperationData/HospitalConsulationOperationDataVo.cs
@@ -51,7 +51,7 @@ namespace Fx.Amiya.Background.Api.Vo.HospitalConsulationOperationData
         /// 新客业绩
         /// </summary>
 
-        public int NewCustomerDealPrice { get; set; }
+        public decimal NewCustomerDealPrice { get; set; }
 
         /// <summary>
         /// 新客客单价
@@ -80,7 +80,7 @@ namespace Fx.Amiya.Background.Api.Vo.HospitalConsulationOperationData
         /// 老客业绩
         /// </summary>
 
-        public int OldCustomerDealPrice { get; set; }
+        public decimal OldCustomerDealPrice { get; set; }
 
         /// <summary>
         /// 老客客单价
